requirements.txt
streamlit
graphviz
langchain-groq
google-generativeai
python-dotenv
os